requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0
scikit-learn>=1.3.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# Data processing
scipy>=1.10.0
geopandas>=0.13.0
rasterio>=1.3.0
shapely>=2.0.0
pyproj>=3.6.0

# API clients
httpx>=0.24.0
requests>=2.31.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-image>=0.20.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0

# Optional: Hyperparameter tuning
optuna>=3.2.0
